Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/script-src-elem-allowed-src-blocked-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-elem'.

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.

Harness Error (TIMEOUT), message = null

NOTRUN Should not fire a security policy violation event
